requirements.txt
mailjet-rest==1.3.4
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
flask-limiter==3.5.0
flask-talisman==1.1.0 